configure el aplication.properties para despliege en railway tambien la configuracion del cross origin con la url de vercel del frontend estoy haciendo pruebas de desplege

+++ src/main/java/com/gestion/automange/service/IUsuarioService.java
@@ -15,7 +15,7 @@ public interface IUsuarioService {
 
 	public void delete(Integer id);
 
-	Optional<Usuario> findById(Long id);
+	Optional<Usuario> findById(Integer id);
 
 	Optional<Usuario> findByEmail(String email);
